requirements.txt
django==5.0
python-telegram-bot==13.15